+++ include/capstone/aarch64.h
@@ -1972,13 +1972,13 @@ typedef union {
 	aarch64_sysreg sysreg;
 	aarch64_tlbi tlbi;
 	aarch64_ic ic;
-	uint64_t raw_val;
+	int raw_val;
 } aarch64_sysop_reg;
 
 typedef union {
 	aarch64_dbnxs dbnxs;
 	aarch64_exactfpimm exactfpimm;
-	uint64_t raw_val;
+	int raw_val;
 } aarch64_sysop_imm;
 
 typedef union {
@@ -1997,7 +1997,7 @@ typedef union {
 	aarch64_bti bti;
 	aarch64_svepredpat svepredpat;
 	aarch64_sveveclenspecifier sveveclenspecifier;
-	uint64_t raw_val;
+	int raw_val;
 } aarch64_sysop_alias;
 
 /// Operand type for instruction's operands
